updated auth.php with comments, adjusted authentication.php to auth.php so authorization functions can be included aswell
