+++ public/view-movie.php
@@ -1,7 +1,7 @@
 <?php
 // imports
+require_once("../utils/auth.php");
 require_once("../utils/database.php");
-require_once("../utils/authentication.php");
 require_once("../utils/filter.php");
 require_once("../utils/movies.php");
 require_once("../utils/functions.php");
